# refactor(record): prefer arrays over sets for lookups of blocked keys