--- a/library/src/schemas/record/record.ts
+++ b/library/src/schemas/record/record.ts
@@ -165,7 +165,7 @@ export function record<
       // Note: `Object.entries(...)` converts each key to a string
       Object.entries(input).forEach(([inputKey, inputValue]) => {
         // Exclude blocked keys to prevent prototype pollutions
-        if (!BLOCKED_KEYS.has(inputKey)) {
+        if (!BLOCKED_KEYS.includes(inputKey)) {
           // Get current path
           const path = getCurrentPath(info, {
             schema: 'record',