## Going from `Set` to `Array`

Following #67, this PR swaps the container of `BLOCKED_KEYS` from a `Set` into an `Array`. This is because although `Set#has` is $O(1)$, the constant coefficient for this is actually quite expensive due to underlying hashing algorithms. In comparison, the $O(n)$ function `Array#includes` should be faster **for small arrays** for the following reasons:

1. `Array#includes` is a built-in method which does not perform any hashing.
2. With its elements being collocated, arrays are typically friendlier to the CPU cache. This is not the case for sets whose elements are sparsely located in a buffer.

In the case of `BLOCKED_KEYS`, three elements are certainly sufficiently small to justify the linear search. See the discussion in https://github.com/fabian-hiller/valibot/pull/67#discussion_r1285232159 for more details and some benchmarks.

## Short-circuiting

Along the way, I also inverted the logic in the `forEach` loops to accommodate for short-circuiting and fewer indentations. Moreover, I converted a `forEach` loop in `record.ts` to a `for...of` loop because the latter is generally considered to be faster.