@fabian-hiller 2023-08-06
Thank you for your contribution. Yes, we can change set to array.

> Along the way, I also inverted the logic in the forEach loops to accommodate for short-circuiting and fewer indentations.

I prefer the way it was before. If something is in a condition, it is quite clear in which case it is executed.

> Moreover, I converted a forEach loop in record.ts to a for...of loop because the latter is generally considered to be faster.

Do you have proof of this? I suspect it doesn't matter for `Object.entries()`. And if it is faster, I would change it not only here, but in the library in general, so that the code is consistent.