@BastiDood 2023-08-06
> I prefer the way it was before. If something is in a condition, it is quite clear in which case it is executed.

Fair enough. This will be reverted ASAP.

> Do you have proof of this? I suspect it doesn't matter for Object.entries(). And if it is faster, I would change it not only here, but in the library in general, so that the code is consistent.

I'll cite [one](https://leanylabs.com/blog/js-forEach-map-reduce-vs-for-for_of/) of many articles in particular, but this is _apparently_ a well known performance pitfall. I'll also go ahead and cite [ThePrimeagen](https://github.com/ThePrimeagen)'s streams, but I can't of course scroll through all the Twitch VODs. Would you like to proceed with this refactor?