@BastiDood 2023-08-06
Performance aside for a moment, I must say that I do prefer the `for...of` syntax over the `forEach` syntax simply because it semantically feels more "side effect-y". Normally, when we invoke methods on an `Array`, we chain together `map` and `filter`, which feels a lot like functional programming. However, `forEach` is a little different in that it goes against that expectation of purity.

So at least for the semantics and aesthetics, I can argue for the `for...of`.

Performance-wise, however, I can't say I'm surprised with the results in many benchmarks. The `forEach` method necessitates the creation of a brand new closure _on top of what would be a `for...of` loop anyway_. This is opposed to the `for...of` loop, which is already a primitive control flow mechanism in and of itself, which _may be easier to optimize_. Of course, this is only an educated guess and speculation. I do hope I got my point across, though.