@BastiDood 2023-08-06
Hello there, I went ahead and implemented the refactor in a184a77. Let me know if you want me to `git revert` the change. 👍

Also, now that #71 has been merged, I rebased this branch on top of the latest `main` so that we can run CI here.