@vijayksingh 2023-08-06
> > Would you like to proceed with this refactor anyway?
> 
> I am not sure yet. I understand the performance advantage. However, I wonder if this approach is faster across the board or only after a certain number of object/array entries.

https://twitter.com/tmikov/status/1688271863087280128

There is some discussion regarding the same here.