@fabian-hiller 2023-08-06
Before we revise the loops, I would prefer a benchmark that demonstrates a performance difference for each scheme. For example, I'm not sure if `for…of` with `.entries()` is really faster.