@BastiDood 2023-08-07
Hello! I am glad to report that after some benchmarking, I have observed considerable improvements in the average execution time.

## Methodology

I created an untracked file named `library/bench.mjs`, which serves as the benchmark. Essentially, it pre-allocates 16K randomly generated objects that contains a number, a string, a date, and a dictionary of strings to numbers. Here, I used 16 key-value pairs for the record, which is admittedly an arbitrary decision.

```js
import { getRandomValues } from 'node:crypto';
import { array, record, object, number, string, date } from './dist/index.js';

function randString() {
    const bytes = getRandomValues(new Uint8Array(21));
    return Buffer.from(bytes).toString('base64');
}

function randDict(n) {
    const dict = {};
    for (let i = 0; i < n; ++i)
        dict[randString()] = Math.random();
    return dict;
}

function randObj(dict) {
    return {
        test: Math.random(),
        name: randString(),
        expires: new Date,
        dict,
    };
}

function randArr(arrLen, dictLen) {
    const objs = Array(arrLen);
    for (let i = 0; i < arrLen; ++i)
        objs[i] = randObj(randDict(dictLen))
    return objs;
}

const ArrayOfObjects = array(object({
    test: number(),
    name: string(),
    expires: date(),
    dict: record(string(), number()),
}));

// Pre-allocate 16K randomly generated objects so that
// we don't incur the overhead during the benchmarking.
const objs = randArr(2 ** 14, 16);

// Warm up the optimization
ArrayOfObjects.parse(objs);

// Do the actual benchmark
const start = performance.now();
ArrayOfObjects.parse(objs);
const end = performance.now();

// Print the results
console.log(end - start);
```

I then ran the following Powershell script to run the benchmark ten times.

```pwsh
cd library

# Before PR
git switch main
pnpm build
for ($i = 0; $i -lt 10; ++$i) { node bench.mjs }

# After PR
git switch prefer-array
pnpm build
for ($i = 0; $i -lt 10; ++$i) { node bench.mjs }
```

## Results

> **Important**
> Here are the relevant hardware specifications of my setup: **Intel® Core™ i7-10750 CPU @ 2.60GHz (12 CPUs)** with **8192 MB of RAM** on **Node.js v20.5.0** for **Windows 10 Build 19045.3271**.

All of the results below are in **milliseconds**.

Run | Before PR | After PR
:-- | --: | --:
1 | 1903.3630999922752 | 1895.031199991703
2 | 1900.3222999572754 | 1894.6241000294685
3 | 1902.9138000011444 | 1898.1823000311852
4 | 1938.1334999799728 | 1885.9580000042915
5 | 1922.8105999827385 | 1897.9362999796867
6 | 1920.4319999814034 | 1892.5608999729156
7 | 1895.236999988556 | 1885.3299000263214
8 | 1928.294000029564 | 1914.0356999635696
9 | 1901.0822999477386 | 1901.5369000434875
10 | 1928.0683000087738 | 1913.57800000906

We thus have the following average execution times in **millseconds**:

Before PR | After PR
:-- | --:
1914.065689986944 | 1897.877330005169

As we can see from the results, we have indeed gone faster by **16 milliseconds** for 16K objects! That's effectively one frame in a 60 FPS render loop. Okay, of course in the real world, we wouldn't really spawn 16K objects in an array all at once, but this at least gives us an idea of the performance improvement nevertheless.