@BastiDood 2023-08-07
> I would be interested to see if there is also a small difference if only 100 or 1,000 objects are used instead of 16K.

I'll be completely honest with you. I did not expect the benchmarks to be this much better across the board. 😅

Elements | Before PR | After PR
:-- | --: | --:
128 | 6.1228900015354 | 6.0720500051975
1024 | 39.009030002356 | 36.450479996204
16384 | 1914.065689986944 | 1897.877330005169

I find the 1024 case to be the most fascinating because the range from 128-1024 sounds like a realistic figure in production (unlike the rather inflated 16K case). It seems that we get the biggest wins here. Or at least I think this is the point where the optimizing compiler really shows its true colors.